fix: replace wrong assertion with if statement
the assertion is invalid because there is more than one way to leave the
:searching loop, other than finding no more matches.
If a multiline match is found that can't be rematched on one line, the
remainder may have been printed. If no match is found after the
search continuation the remainder is still printed and the assertion
triggers.
The same goes for the loop that prints the after context lines.